mouse_battery_hwinfo_requirements.txt
pywinusb == 0.4.2
pywin32 == 306
Pillow == 10.1.0
pystray == 0.19.5